backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
cryptography>=42.0.0

# Build
pyinstaller>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
